pyproject.toml
[project]
name = "market-db"
version = "0.1.0"
description = "DuckDB warehouse for DataBento CME MDP 3.0 (options & futures)"
requires-python = ">=3.10"
dependencies = ["duckdb>=1.0.0","pyyaml>=6.0","typer>=0.12.3","python-dotenv>=1.0.1"]

[project.scripts]
market-db = "orchestrator:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


